Fix creation time of docker network in integration tests
Fixes: #21931
